Scheduling fixing

+++ student-distrib/term_switch.h
@@ -18,7 +18,7 @@ typedef struct __attribute__((packed)) term{
 	int buf_count;
     uint8_t screenx;
     uint8_t screeny;
-	uint8_t curr_pid;
+	uint32_t curr_pid;
     uint32_t esp;
     uint32_t ebp;
